Proposed File link fixes
Fix the recognition of file links in HtmlEditorField.js (regex was lacking the ',')

Rewrite old file links (url) to new file_link shortcode HtmlEditorField.php .
As the old url format for file links is no longer supported by HtmlEditorField, this allows for a simple upgrade from earlier versions.
A migration task might be better suited for this, but there are quite some references to the older file link format left in the code.